% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeteroPCA.R
\name{Heteropca}
\alias{Heteropca}
\title{A method of PCA}
\usage{
Heteropca(hatsigma, r, T)
}
\arguments{
\item{hatsigma}{the matrix}

\item{r}{the number of factors to be calculated}

\item{T}{max times bound of the iteration}
}
\value{
a random sample of size \code{n}
}
\description{
A method of PCA
}
\examples{
\dontrun{
a <- matrix(1:9,nrow=3)
Heteropca(a,2,100)
}
}
