% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robpre.R
\name{robpre}
\alias{robpre}
\title{A method of figure out precision matrix}
\usage{
robpre(sigmahat, rho0)
}
\arguments{
\item{sigmahat}{the matrix to be inverse}

\item{rho0}{the radius}
}
\value{
a list of precision matrix,parameter gamma, and eigens of precision matrix
}
\description{
A method of figure out precision matrix
}
\examples{
\dontrun{
#a <- matrix(c(2,0,0,3),nrow=2)
b <- robpre(a,0.002)
s <- svd(b)
s$u
s$d
}
}
