\name{StatComp21013-package}
\alias{StatComp21013-package}
\alias{StatComp21013}
\docType{package}
\title{
  Final project
}
\description{
  Homework and 2 functions.
}
\details{
  write 2 fuctions of thesis below.
}
\author{
Hu Zongqing

Hu Zongqing <617721320@qq.com>
}
\references{
  Nguyen V A, Kuhn D, Mohajerin Esfahani P. Distributionally robust inverse covariance estimation: The Wasserstein shrinkage estimator[J]. Operations Research, 2021.
  Zhang A R, Cai T T, Wu Y. Heteroskedastic PCA: Algorithm, optimality, and applications[J]. arXiv preprint arXiv:1810.08316, 2018.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
